@@ -1,31 +1,25 @@
 {
   "name": "miro-team-topologies",
-  "version": "0.0.2",
-  "private": true,
+  "version": "2.0.0",
+  "licence": "MIT",
   "scripts": {
-    "build": "webpack --mode production",
-    "dev": "webpack --mode development",
-    "watch": "webpack -w --mode development"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@types/react": "16.9.49",
-    "@types/react-dom": "16.9.8",
-    "css-loader": "4.3.0",
-    "file-loader": "6.1.0",
-    "react-inlinesvg": "2.1.0",
-    "style-loader": "1.2.1",
-    "svg-inline-loader": "0.8.0",
-    "ts-loader": "8.0.3",
-    "typescript": "4.0.2",
-    "url-loader": "4.1.0",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12"
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
   },
   "dependencies": {
-    "copy-to-clipboard": "3.2.0",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-markdown": "^5.0.3"
+    "mirotone": "3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-inlinesvg": "^3.0.0",
+    "react-markdown": "^8.0.3"
+  },
+  "devDependencies": {
+    "@mirohq/websdk-types": "^2.3.0",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
+    "@vitejs/plugin-react-refresh": "1.3.6",
+    "typescript": "4.7.4",
+    "vite": "3.0.3"
   }
 }
