@@ -1,3 +1,5 @@
+import { FontFamily, TextAlign, TextAlignVertical, ShapeType } from '@mirohq/websdk-types';
+
 export enum TEAM_ENUM {
   StreamAligned = 0,
   Platform = 1,
@@ -7,33 +9,23 @@ export enum TEAM_ENUM {
   Collaboration = 5,
   Facilitating = 6,
   Xaas = 7,
-  FlowOfChange = 8,
+  FlowOfChange = 8
 }
 export interface TeamElementStyle {
-  shapeType: number
-  backgroundColor: string
-  backgroundOpacity: number
-  borderColor: string
-  borderWidth: number
-  borderOpacity: number
-  borderStyle: number
-  fontFamily: number
-  textColor: string
-  textAlign: string
-  textAlignVertical: string
-  fontSize: number
-  bold: number
-  italic: number
-  underline: number
-  strike: number
-  highlighting: string
+  fillColor?: string | undefined;
+  fontFamily?: FontFamily | undefined;
+  fontSize?: number | undefined;
+  textAlign?: TextAlign | undefined;
+  textAlignVertical?: TextAlignVertical | undefined;
+  shapeType: ShapeType;
 }
+
 export interface TeamElement {
-  getTeamEnum(): TEAM_ENUM
-  getStyle(): TeamElementStyle
-  getName(): string
-  getShapeSize(): {height: number; width: number}
-  getPreview(): string
-  getIcon(): string
-  getClassName(): string
+  getTeamEnum(): TEAM_ENUM;
+  getStyle(): TeamElementStyle;
+  getName(): string;
+  getShapeSize(): { height: number; width: number };
+  getPreview(): string;
+  getIcon(): string;
+  getClassName(): string;
 }
